+++ CHANGELOG.md
@@ -7,7 +7,23 @@ This page shows a detailed overview of the changes between versions without the 
 	## __WORK IN PROGRESS__
 -->
 
-## __WORK IN PROGRESS__
+## 0.5.0 (2026-03-05)
+- Feature: Allows searching for thread nodes in the network graph by extended address
+- Adjustment: Also sorts neighbors in network graph details by quality
+- Fix: Fixed datatype of some custom cluster attributes
+- Fix: Only allows starting a node update when the node is connected/subscribed currently
+- Fix/Enhancement: Update matter.js to 0.17.0-nightly
+    - Rework and further optimize Discovery and connection logic
+    - Fix cases where BLE disconnects could crash the server 
+
+## 0.4.3 (2026-02-25)
+- Adjustment: Sort connection quality entries in network graph details by quality
+- Fix: Fixed encoding of custom array-based data types
+- Fix: Triggers a full node-update when a (Bridged Node) Basic Information cluster attribute changes
+- Fix: Uses the same kind of signal icons for all signal levels
+- Fix/Enhancement: Update matter.js to 0.16.11-nightly
+
+## 0.4.2 (2026-02-22)
 - Fix: Sanitize wifi/thread credentials in the log when setting them
 - Fix: Update matter.js to 0.16.10